Early aggregation of lookup tables
