+++ Physics/ItoKMC/PlasmaModels/ItoKMCAir3LFA/CD_ItoKMCAir3LFA.H
@@ -14,7 +14,7 @@
 
 // Our includes
 #include <CD_ItoKMCPhysics.H>
-#include <CD_LookupTable1D.H>
+#include <CD_LookupTable.H>
 #include <CD_NamespaceHeader.H>
 
 namespace Physics {
