+++ Source/Utilities/CD_LookupTable1D.H
@@ -23,37 +23,7 @@
 
 // Our includes
 #include <CD_NamespaceHeader.H>
-
-/*!
-  @brief Strategy for obtaining data when requesting data that is out of range. 
-  We can use either constant (i.e. virtually padding the data with the first/last entry) or 
-  with linear interpolation (i.e. virtually padding the data with a slope connecting the first
-  two or last two entries in the table). 
-*/
-enum class OutOfRangeStrategy
-{
-  Constant,
-  Linear
-};
-
-/*!
-  @brief For identifying columns in the LookupTable1D.
-*/
-enum StateStatus
-{
-  Column  = 0,
-  Sorted  = 1,
-  Uniform = 2,
-};
-
-/*!
-  @brief Enum for classifying how table is spaced. 
-*/
-enum class TableSpacing
-{
-  Uniform,
-  Exponential
-};
+#include <CD_LookupTable.H>
 
 /*!
   @brief Class for looking up and interpolation (x,y) data.
